feat(gl): serve OpenGL ES 1.1 and ES 2 from one DrawList backend
gles2.rs becomes gl/mod.rs — the generation-independent part: the DrawList
walk, the image and font-atlas caches, batching by texture and scissor, and
the physical clip arithmetic. What actually differs between GL generations
moves into gl/es2.rs (the shader program Symbian uses) and gl/es1.rs (a new
fixed-function pipeline), which satisfy one interface: new, destroy,
begin_frame, set_blend, bind_vertices, unbind_vertices. Exactly one compiles,
chosen by the new `gles1` feature.

The ES 2 shaders have exact fixed-function equivalents, which is why this is
a substitution rather than a rewrite: the vertex shader's pixel-to-NDC
transform including its Y flip is glOrthof(0, w, h, 0, -1, 1), and
`texture2D(u_texture, v_uv) * v_color` is GL_MODULATE over a per-vertex color
array. The interleaved vertex buffer is reused byte for byte — same offsets
0/8/16, same stride.

PowerVR MBX Lite exposes no GL_OES_blend_func_separate, so the ES 1.1 path
uses plain glBlendFunc. That is sound here because the drawable is opaque and
nothing reads its alpha back, and the reasoning is recorded next to the call.

Also widens 8 GL cfg gates and 3 fallback arms in lib.rs from
target_os = "none" to admit feature = "bare-platform". Without this the GL
module was compiled out of existence for armv6-apple-ios while the build,
the link and the archive all still succeeded — the same trap the no_std
switch hit, failing silently a second time.

All ten backend tests pass on both pipelines; host-ES2, host-ES1 and
armv6-ES1 all build with zero warnings.
