+++ engine/symbian/src/lib.rs
@@ -31,8 +31,8 @@ use pocketjs_core::raster;
 use pocketjs_core::Ui;
 
 pub mod extension;
-#[cfg(any(target_os = "none", test))]
-mod gles2;
+#[cfg(any(target_os = "none", feature = "bare-platform", test))]
+mod gl;
 
 #[cfg(any(target_os = "none", feature = "bare-platform", test))]
 const C_MALLOC_ALIGNMENT: usize = 8;
@@ -148,11 +148,11 @@ fn clear_framebuffer() {
 /// Reset the single UI instance. `raster_density == 0` selects density 1.
 #[no_mangle]
 pub extern "C" fn ui_init(raster_density: u32) {
-    #[cfg(target_os = "none")]
+    #[cfg(any(target_os = "none", feature = "bare-platform"))]
     unsafe {
         // This call may happen without a current GL context, so the backend
         // only marks its caches stale and defers replacement until render.
-        gles2::invalidate_resources();
+        gl::invalidate_resources();
     }
     unsafe {
         UI = Some(Ui::new_with_raster_density(raster_density.max(1)));
@@ -163,9 +163,9 @@ pub extern "C" fn ui_init(raster_density: u32) {
 /// Drop all retained UI, texture, font, and framebuffer allocations.
 #[no_mangle]
 pub extern "C" fn ui_shutdown() {
-    #[cfg(target_os = "none")]
+    #[cfg(any(target_os = "none", feature = "bare-platform"))]
     unsafe {
-        gles2::invalidate_resources();
+        gl::invalidate_resources();
     }
     unsafe {
         UI = None;
@@ -373,13 +373,13 @@ pub extern "C" fn ui_load_styles(ptr: *const u8, len: usize) -> i32 {
 pub extern "C" fn ui_load_font_atlas(ptr: *const u8, len: usize) -> i32 {
     let blob = unsafe { bytes(ptr, len) };
     let loaded = ui().load_font_atlas(blob);
-    #[cfg(target_os = "none")]
+    #[cfg(any(target_os = "none", feature = "bare-platform"))]
     if loaded {
         if let Some(&slot) = blob.get(12) {
             unsafe {
                 // Loading happens in a host callback, not necessarily with
                 // QGLWidget's context current. Defer GL deletion/re-upload.
-                gles2::invalidate_font(slot);
+                gl::invalidate_font(slot);
             }
         }
     }
@@ -400,27 +400,27 @@ pub extern "C" fn ui_tick() {
 
 #[no_mangle]
 pub extern "C" fn ui_gl_initialize() -> i32 {
-    #[cfg(target_os = "none")]
+    #[cfg(any(target_os = "none", feature = "bare-platform"))]
     unsafe {
-        return gles2::initialize() as i32;
+        return gl::initialize() as i32;
     }
-    #[cfg(not(target_os = "none"))]
+    #[cfg(not(any(target_os = "none", feature = "bare-platform")))]
     0
 }
 
 #[no_mangle]
 pub extern "C" fn ui_gl_reset_resources() {
-    #[cfg(target_os = "none")]
+    #[cfg(any(target_os = "none", feature = "bare-platform"))]
     unsafe {
-        gles2::reset_resources();
+        gl::reset_resources();
     }
 }
 
 #[no_mangle]
 pub extern "C" fn ui_gl_shutdown() {
-    #[cfg(target_os = "none")]
+    #[cfg(any(target_os = "none", feature = "bare-platform"))]
     unsafe {
-        gles2::shutdown();
+        gl::shutdown();
     }
 }
 
@@ -433,9 +433,9 @@ pub extern "C" fn ui_gl_render(
     window_width: i32,
     window_height: i32,
 ) -> i32 {
-    #[cfg(target_os = "none")]
+    #[cfg(any(target_os = "none", feature = "bare-platform"))]
     unsafe {
-        return gles2::render(
+        return gl::render(
             ui(),
             target_x,
             target_y,
@@ -445,7 +445,7 @@ pub extern "C" fn ui_gl_render(
             window_height,
         ) as i32;
     }
-    #[cfg(not(target_os = "none"))]
+    #[cfg(not(any(target_os = "none", feature = "bare-platform")))]
     {
         let _ = (
             target_x,
@@ -471,9 +471,9 @@ pub extern "C" fn ui_gl_render_over(
     window_width: i32,
     window_height: i32,
 ) -> i32 {
-    #[cfg(target_os = "none")]
+    #[cfg(any(target_os = "none", feature = "bare-platform"))]
     unsafe {
-        return gles2::render_over(
+        return gl::render_over(
             ui(),
             target_x,
             target_y,
@@ -483,7 +483,7 @@ pub extern "C" fn ui_gl_render_over(
             window_height,
         ) as i32;
     }
-    #[cfg(not(target_os = "none"))]
+    #[cfg(not(any(target_os = "none", feature = "bare-platform")))]
     {
         let _ = (
             target_x,
